requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
psutil==5.9.5
